Lesson 3 (B - TapeEquilibrium)
Iterating through array and slicing it up from current index till end. Comparing sum of each slice and finding the minimum. Note: It's not passing time complexity, slicing array + summing up the items in it is slow, this can be separated into "smarted" way with just adding value from next item that is added in iteration.