Lesson 5 (A - PassingCars)
This solution should be viewed from a perspective of a car moving W->E (0). We count all the incoming cars (sum) and then iterate through array, if it's 0 we encounter a 0 that means that it will pass all the rest of the cars that are coming its way (1), and if we encounter 1 that means that car has passed and we should remove it from incoming passing cars.