Lesson 5 (C - GenomicRangeQuery)
Solution is correct but it's super slow. Whole algorithm is not optimised at all. Slicing array and find minimal each time is super time consuming. My initial idea is to map indexes where each 1, 2, 3, and 4 values are. And then find a smart way of finding if 1 is in index start/end defined in P/Q