Lesson 6 (D - NumberOfDiscIntersections)
Works OK but it's too slow. Idea is to map all discs to coordinate system where we check if disk is touching the X value of coordinate system. So for each disk we create `Set` of X-coordinates. Then we iterate through all discs and check with which other discs (all after current disk index (to avoid counting twice) have the same X values.